% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_ordination.R
\name{apply_ordination}
\alias{apply_ordination}
\title{Apply ordination procedures for multivariate statistical analysis}
\usage{
apply_ordination(data, protocol = "1", dimensions = 2,
  exception_columns = NULL, variable_tags = NULL, coda_override = NULL,
  coda_transformation_method = "CLR", coda_alr_base = 1,
  coda_pca_method = "robust", init_seed = 0, coda_samples = 100)
}
\arguments{
\item{data}{Data frame, including compositional and petrographic data.}

\item{protocol}{Character, cerUB protocol to be applied.
"1": Analysis of compositional data;
"2a": Analysis of petrographic data (relative ranking difference);
"2b": Analysis of petrographic data (neighbor interchange);
"3": Analysis of compositional data and petrographic data
(relative ranking);
"4": Analysis of compositional data and petrographic data
(relative ranking) to characterize provenance.}

\item{dimensions}{Numeric, number of dimensions of the ordination object.}

\item{exception_columns}{Numeric, the vector of variables names to be
searched for exceptions.}

\item{variable_tags}{Character, two-column data frame containing (1)
the names of variables and (2) their tags.}

\item{coda_override}{Character, vector with the names of the
compositional variables.}

\item{coda_transformation_method}{Character, the log-ratio transformation
     to be applied:
"ALR" for additive log-ratio,
"CLR" for centered log-ratio,
"ILR" for isometric log-ratio.
Additionally, accepts "log" for applying
logarithmic transformation and "std" for standardization
(scaled and centred).}

\item{coda_alr_base}{Character/Numeric, the name/index of the variable
to be used as divisor in additional log-ratio
transformation.}

\item{coda_pca_method}{Character, Principal Components Analysis (PCA) method:
"standard" for standard PCA, "robust" for robust PCA.}

\item{init_seed, coda_samples}{Numeric, arguments passed to
\code{\link[cerUB]{princomp_coda}}.}
}
\value{
Ordination object containing the projection of observations (scores)
and variables (loadings) in 'n' dimensions, the distance matrix used
(dist_matrix), and an approximation of the fitness of projections.
}
\description{
#' Applies a given ordination procedure to a data set
and returns a ordination object.
}
