% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princomp_coda.R
\name{best_pcaCoDa}
\alias{best_pcaCoDa}
\title{Find the best 'pcaCoDa' (robCompositions)}
\usage{
best_pcaCoDa(dt, method = "robust", init_seed = 0, samples = 100)
}
\arguments{
\item{dt}{Data frame containing compositional data}

\item{method}{Character, "standard" for standard PCA, "robust" for robust PCA.}

\item{init_seed}{Numeric, the seed for the random number generator used in
\code{\link{best_pcaCoDa}}).}

\item{samples}{Numeric, the number of iterations applying to samples in
\code{\link{best_pcaCoDa}}) and
maxiter in \code{\link[pcaPP]{PCAgrid}}).}
}
\description{
Searches the best projection given by \code{\link[robCompositions]{pcaCoDa}}.
}
