% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_variables.R
\name{code_variables}
\alias{code_variables}
\title{Codify petrographic variable names}
\usage{
code_variables(data)
}
\arguments{
\item{data}{A petrographic data frame
using cerUB naming system.}
}
\description{
Creates a code for each variable in a petrographic data frame
using cerUB naming system, and return a two-column data frame
relating original names with their codes.
}
