% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers_per_group}
\alias{detect_outliers_per_group}
\title{Detect outliers per group}
\usage{
detect_outliers_per_group(distMatrix, groups, method = "MD",
  criterium = "MAD", LOF_k = 2, MAD_trim = 2, boxplot_trim = 1.5)
}
\arguments{
\item{distMatrix}{Numeric, distance matrix}

\item{groups}{Factor, vector containing the assignation of each observation
in the matrix to a specific group.}

\item{method}{Character, method for measuring separation of one point
from all other points.
The following are accepted:
"MdD": median distance;
"MD": average (mean) distance;
"MAH": Mahalanobis distances (\code{\link[robCompositions]{outCoDa}});
"LOF": Local Outlier Factor Score (\code{\link[dbscan]{lof}}).}

\item{criterium}{Numeric/Character, the criterium used for separating outliers.
The following are accepted:
<Numeric, 0-1>: number between 0 and 1, sets a quantile type of threshold;
"boxplot": outliers are those singled out as points in a boxplot;
"MAD": threshold is given by Median Absolute Deviation.}

\item{LOF_k}{Numeric, when method = "LOF", the size of the neighborhood.
See \code{\link[dbscan]{lof}}.}

\item{MAD_trim}{Numeric, when criterium = "MAD", the multipler of MAD
to calculate a outlier threshold.}

\item{boxplot_trim}{Numeric, when criterium = "boxplot", the multipler of
the interquartile range (IQR) to calculate a outlier
threshold.}
}
\description{
Detects outliers per each group in a distance matrix using a certain method and
following a certain criterium.
}
\examples{

\dontrun{

pca <- princomp(iris[, 1:4])

irisSpeciesOutliers_MD <- detect_outliers_per_group(dist(iris[, 1:4]),
                                                    iris$Species,
                                                    method = "MD",
                                                    criterium = "MAD")
irisSpeciesOutliers_LOF <- detect_outliers_per_group(dist(iris[, 1:4]),
                                                     iris$Species,
                                                     method = "LOF",
                                                     criterium = "MAD")
plot(pca$scores[, 1:2],
    col = iris$Species,
    main = "Outliers per group")
points(pca$scores[irisOutliers_MD$index, 1:2],
       col = "purple", pch = 2, cex = 1.5)
points(pca$scores[irisOutliers_LOF$index, 1:2],
       col = "orange", pch = 6, cex = 1.5)
legend(0.65 * max(pca$scores[,1]), max(pca$scores[,2]),
       c("MD", "LOF"), pch = c(2, 6), col = c("purple", "orange"))

}

}
