% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_gower.R
\name{extended_gower}
\alias{extended_gower}
\title{Extended Gower with exceptions}
\usage{
extended_gower(data, variable_sets, method = "RRD",
  exception_columns = NULL, exception_values = NULL,
  exception_distance = 0)
}
\arguments{
\item{data}{Data frame containing ordinal variables and optionally
compositional data (log-ratios).}

\item{variable_sets}{List containing vectors with the numeric index of variables
of different kind.}

\item{method}{Character, "NI" for neighbour interexchange,
"RRD" for relative ranking difference,
"MM" for mixed mode.}

\item{exception_columns}{Numeric, the vector of variables names to be
searched for exceptions.}

\item{exception_values}{Numeric, the value to be set with distance.}

\item{exception_distance}{Numeric, the index of the variable to be used as
divisor in additive log-ratio transformation.}
}
\description{
Calculate the extended Gower coefficient for archaeometric data sets,
considering exceptional values in ordinal variables, using a modified version
of \code{\link[ade4]{dist.ktab}}.
}
