% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_coda}
\alias{get_coda}
\title{Get indexes of compositional variables}
\usage{
get_coda(data, coda_variables, transformation_method = "")
}
\arguments{
\item{data}{Data frame containing petrographic variables that use the
cerUB naming system.}

\item{coda_variables}{Character/Numeric, vector with the names/indexes of the compositional variables.}

\item{transformation_method}{Character, vector specifying a transformation:
"ALR" -> additive log-ratio,
"CLR" -> centered log-ratio,
"ILR" -> isometric log-ratio.
Additionally, accepts "log" for applying
logarithmic transformation.}
}
\description{
Get indexes of compositional variables that were transformed using
\code{\link[cerUB]{transform_coda}}.
}
