% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_petro}
\alias{get_petro}
\title{Get indexes of petrographic variables}
\usage{
get_petro(data)
}
\arguments{
\item{data}{Data frame containing petrographic variables that use the
cerUB naming system.}
}
\description{
Get indexes of petrographic variables that use the cerUB naming system.
}
