% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_provenance}
\alias{get_provenance}
\title{Get indexes of provenance-related variables}
\usage{
get_provenance(data, coda_variables, transformation_method = "")
}
\arguments{
\item{data}{Data frame containing petrographic variables that use the
cerUB naming system.}

\item{coda_variables}{Character/Numeric, vector with the names/indexes of
the compositional variables.}

\item{transformation_method}{Character, vector specifying a transformation:
"ALR" -> additive log-ratio,
"CLR" -> centered log-ratio,
"ILR" -> isometric log-ratio.
Additionally, accepts "log" for applying
logarithmic transformation.}
}
\value{
List of two numeric vectors containing the column indexes of (1)
petrographic and (2) compositional variables.
}
\description{
Get indexes of the compositional and petrographic variables indicative
of provenance. Petrographic variables must use the cerUB naming system.
Compositional variables must have been transformed using
\code{\link[cerUB]{transform_coda}}.
}
