% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds.R
\name{nmds}
\alias{nmds}
\title{Non-metric Multidimensional Scaling (NMDS)}
\usage{
nmds(distance_matrix, original_data, variable_tags = c(), dimensions = 2,
  init_seed = 0, trymax = 100, autotransform = FALSE)
}
\arguments{
\item{distance_matrix}{distance or dissimilarity matrix}

\item{original_data}{data frame containing the original data}

\item{variable_tags}{Character, two-column data frame containing (1)
the names of variables and (2) their tags.}

\item{dimensions}{Numeric, number of dimensions of the projection equivalent to
k in \code{\link[vegan]{metaMDS}}.}

\item{init_seed}{Numeric, the seed for the random number generator
used by \code{\link[vegan]{metaMDS}}.}

\item{trymax, autotransform}{Numeric, Maximum number of random starts in search of
stable solution. Logical, whether to use simple heuristics for
possible data transformation of typical community data (see below).
If you do not have community data, you should probably set
autotransform = FALSE.
Arguments passed to \code{\link[vegan]{metaMDS}}.}
}
\description{
Apply Non-metric Multidimensional Scaling to a given distance matrix,
calculate variable covariances, and the percent of variance explained by
2D and 3D projections.
}
