% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_petro.R
\name{order_petro}
\alias{order_petro}
\title{Order petrographic ordinal variables}
\usage{
order_petro(data)
}
\arguments{
\item{data}{Data frame containing petrographic ordinal variables that use the cerUB naming system.}
}
\description{
Order the values (levels) of petrographic ordinal variables (factors) that use the cerUB naming system.
}
