% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa.R
\name{pcoa}
\alias{pcoa}
\title{Principal Coordinates Analysis}
\usage{
pcoa(distance_matrix, original_data, variable_tags = NULL, dimensions = 2)
}
\arguments{
\item{distance_matrix}{distance or dissimilarity matrix}

\item{original_data}{data frame containing the original data}

\item{variable_tags}{Character, two-column data frame containing (1)
the names of variables and (2) their tags.}

\item{dimensions}{Numeric, number of dimensions of the projection equivalent to
k in \code{\link[stats]{cmdscale}}}
}
\description{
Apply Principal Coordinates Analysis to a given distance matrix,
calculate variable covariances, and the percent of variance explained by
2D and 3D projections.
}
