% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/princomp_coda.R
\name{princomp_coda}
\alias{princomp_coda}
\title{Principal Components Analysis for compositional data}
\usage{
princomp_coda(dt, transformation_method = "ILR", method = "robust",
  init_seed = 0, samples = 100, alr_base = 1)
}
\arguments{
\item{dt}{Data frame containing compositional data}

\item{transformation_method}{Character, the log-ratio transformation to be applied.
"ALR" -> additive log-ratio,
"CLR" -> centered log-ratio,
"ILR" -> isometric log-ratio.
Additionally, accepts "log" for applying
logarithmic transformation and "std" for standardization
(scaled and centred).}

\item{method}{Character, "standard" for standard PCA, "robust" for robust PCA.}

\item{init_seed}{Numeric, the seed for the random number generator used in
\code{\link{best_pcaCoDa}}).}

\item{samples}{Numeric, the number of iterations applying to samples in
\code{\link{best_pcaCoDa}}) and
maxiter in \code{\link[pcaPP]{PCAgrid}}).}

\item{alr_base}{Character/Numeric, the name/index of the variable to be used
as divisor in additional log-ratio transformation.}
}
\description{
Principal Components Analysis (PCA) of compositional data
after applying log-ratio transformation.
}
