% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replaces NA values}
\usage{
replace_na(x, as_na = c(NULL), method = NULL)
}
\arguments{
\item{x}{Numeric or Character/factor vector.}

\item{as_na}{Numeric or Character, specifies values that may be considered as NA.}

\item{method}{Character, the method used for choosing the replacement ("NULL", "random", "mode", "normal"). Accepts numeric value if x is numeric. Method must be compatible with x class.}
}
\description{
Replaces NA values in a vector according to a given method.
Specific non-NA values can be considered as NA, if given in as_na.
}
\examples{

\dontrun{

replace_na(c(1, 2, NA, 4, 5, 6), method = 3)
replace_na(c(1, 3.5, 9, 2.8, 5.6, 10.4, 0.7, 2.4, 5.5, NA), method = "normal")
replace_na(c(1, 3.5, 9, 2.8, 5.6, 10.4, 0.7, 2.4, 5.5, NA), method = "random")
replace_na(c("A", "B", "A", "F", "K", "B", "O", "A"), method = "mode")

}

}
