% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_coda_names.R
\name{simplify_coda_names}
\alias{simplify_coda_names}
\title{Simplify transformed Compositional data names in ordination object}
\usage{
simplify_coda_names(ordination_object)
}
\arguments{
\item{ordination_object}{Ordination object, as generated by
\code{\link[cerUB]{apply_ordination}}, containing a
"loadings" data frame with transformed CoDa variables.}
}
\description{
Replace composite name of type "transformationMethod-component",
such as CLR-Fe2O3, with a shorter version, such as "Fe2O3".
}
