% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coda.R
\name{transform_coda}
\alias{transform_coda}
\title{Transform compositional data}
\usage{
transform_coda(data, coda_variables, method = c("CLR"), alr_base = 1,
  raw_filename = NULL, trans_filename = NULL, final_filename = NULL)
}
\arguments{
\item{data}{Data frame containing compositional data.}

\item{coda_variables}{Numeric/Character, vector containing the names/indexes
of the compositional variables.}

\item{method}{Character, the log-ratio transformation to be applied.
"ALR" -> additive log-ratio,
"CLR" -> centered log-ratio,
"ILR" -> isometric log-ratio.
Additionally, accepts "log" for applying
logarithmic transformation and "std" for standardization
(scaled and centred).}

\item{alr_base}{Character/Numeric, the name/index of the variable to be used
as divisor in additional log-ratio transformation.
in additive log-ratio transformation.}

\item{raw_filename, trans_filename, final_filename}{Character, file names for
saving the raw (complete), tranformed (only coda), and
final (complete) data sets.}
}
\description{
Transform compositional data in a given data frame and replace it.
}
